[aric-ros] Factor out JointTrajectoryClientWrapperConfig
- to use it from RosControlActionClient
